refactored imports/functions and -d npm dependency
